refactor: SegmentPropertiesTableView
