+++ test-acceptance/src/org/omegat/gui/properties/SegmentPropertiesAreaTest.java
@@ -1,0 +1,53 @@
+/**************************************************************************
+ OmegaT - Computer Assisted Translation (CAT) tool
+          with fuzzy matching, translation memory, keyword search,
+          glossaries, and translation leveraging into updated projects.
+
+ Copyright (C) 2025 Hiroshi Miura
+               Home page: https://www.omegat.org/
+               Support center: https://omegat.org/support
+
+ This file is part of OmegaT.
+
+ OmegaT is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ OmegaT is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ **************************************************************************/
+package org.omegat.gui.properties;
+
+import org.assertj.swing.data.TableCellInSelectedRow;
+import org.junit.Rule;
+import org.junit.Test;
+import org.omegat.gui.main.TestCoreGUI;
+import org.omegat.util.LocaleRule;
+
+import java.util.Locale;
+
+public class SegmentPropertiesAreaTest extends TestCoreGUI {
+
+    private static final String PROJECT_PATH = "test-acceptance/data/project/";
+
+    @Rule
+    public final LocaleRule localeRule = new LocaleRule(new Locale("en"));
+
+    @Test
+    public void testSegmentPropertiesPaneExist() throws Exception {
+        // load project
+        openSampleProject(PROJECT_PATH);
+        robot().waitForIdle();
+        // check a segment properties pane
+        window.scrollPane("Segment Properties").requireEnabled();
+        window.scrollPane("Segment Properties").requireVisible();
+        window.table("SegmentPropertiesTable").requireVisible();
+    }
+
+}
